[Feat](#2) 공통 엔티티 추가 - 식별자 등록
