+++ src/common/entities/common-id.entity.ts
@@ -1,0 +1,10 @@
+import { PrimaryGeneratedColumn } from 'typeorm';
+
+import { CommonEntity } from './common.entity';
+
+export abstract class CommonIdEntity extends CommonEntity {
+  @PrimaryGeneratedColumn({
+    unsigned: true,
+  })
+  id?: number;
+}
