+++ src/common/entities/common.entity.ts
@@ -1,0 +1,9 @@
+import { CreateDateColumn, UpdateDateColumn } from 'typeorm';
+
+export abstract class CommonEntity {
+  @CreateDateColumn()
+  createdAt?: Date;
+
+  @UpdateDateColumn()
+  updatedAt?: Date;
+}
