+++ src/common/entities/index.ts
@@ -1,0 +1,2 @@
+export * from './common-id.entity';
+export * from './common.entity';
