--- src/common/entity/index.ts
@@ -1,1 +1,0 @@
-export * from './common.entity';
