+++ src/common/index.ts
@@ -1,4 +1,4 @@
-export * from './entity';
+export * from './entities';
 export * from './exception';
 export * from './interceptor';
 export * from './interface';
